requirements.txt
fastapi
uvicorn
python-dotenv
langchain
pinecone-client
asyncio
openai
chromadb
PyPDF2
tiktoken
langchain-community
langchain-groq
langchainhub


numpy==1.24.0
opencv-python==4.8.0.76
scikit-learn==1.3.2


# add any other packages you know your project uses
